+++ minecraft-common/src/main/java/com/sigmundgranaas/forgero/minecraft/common/item/DynamicAttributeItem.java
@@ -58,6 +58,8 @@ public interface DynamicAttributeItem extends DynamicAttributeTool, DynamicDurab
 
     UUID ADDITION_LUCK_MODIFIER_ID = UUID.fromString("CC3F55D5-755C-4F38-A497-9C13A33DB5CF");
 
+    UUID ADDITION_ARMOR_MODIFIER_ID = UUID.fromString("AC3F55D5-755C-4F38-A497-9C13A63DB5CF");
+
     PropertyContainer dynamicProperties(ItemStack stack);
 
     PropertyContainer defaultProperties();
@@ -124,6 +126,9 @@ public interface DynamicAttributeItem extends DynamicAttributeTool, DynamicDurab
         int luck = LuckHandler.of(dynamicProperties(stack)).map(Attribute::asInt).orElse(0);
         builder.put(EntityAttributes.GENERIC_LUCK, new EntityAttributeModifier(ADDITION_LUCK_MODIFIER_ID, "Luck addition", luck, EntityAttributeModifier.Operation.ADDITION));
 
+        float armor = Armor.of(dynamicProperties(stack)).asFloat();
+        builder.put(EntityAttributes.GENERIC_ARMOR, new EntityAttributeModifier(ADDITION_ARMOR_MODIFIER_ID, "Armor addition", armor, EntityAttributeModifier.Operation.ADDITION));
+        
         //Attack speed
         float baseAttackSpeed = AttackSpeed.apply(dynamicProperties(stack), target);
         float currentAttackSpeed = AttackSpeed.apply(defaultProperties());